+++ src/cloudwatch-mcp-server/CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## [0.0.5] - 2025-10-06
+
+### Added
+
+- Added tool to analyze CloudWatch Metric data
+
+### Changed
+
+- Updated Alarm recommendation tool to support CloudWatch Anomaly Detection Alarms
+
 ## [0.0.4] - 2025-07-11
 
 ### Changed